requirements.txt
Django>=4.2
psycopg2-binary>=2.9
python-keycloak>=3.9
PyJWT>=2.8
PyPDF2>=3.0
reportlab>=4.0
requests>=2.31
